src/grab.rs: revise code due to config extension
