+++ src/glob.rs
@@ -194,7 +194,6 @@ pub(crate) static RX_LOADING_WORDS: Lazy<Regex> = Lazy::new(|| {
 pub(crate) static RX_SHARE_ELEMENTS: Lazy<Regex> =
     Lazy::new(|| Regex::new(r#"(?i)(\b|_)(share|sharedaddy)(\b|_)"#).unwrap());
 
-pub(crate) static DEFAULT_N_TOP_CANDIDATES: usize = 5;
 pub(crate) static MINIMUM_TOP_CANDIDATES: usize = 3;
 pub(crate) static DEFAULT_CHAR_THRESHOLD: usize = 500;
 pub(crate) static SCORE_ATTR: &str = "data-readability-score";
