+++ src/grab.rs
@@ -13,6 +13,261 @@ use crate::score::*;
 use crate::helpers::*;
 use crate::prep_article::prep_article;
 use crate::Metadata;
+use crate::Readability;
+
+impl Readability {
+    pub fn grab_article(&self, metadata: &mut Metadata) -> Option<Document> {
+        let mut flags =
+            GrabFlags::CleanConditionally | GrabFlags::StripUnlikelys | GrabFlags::WeightClasses;
+    
+        let mut attempts = vec![];
+        loop {
+            let mut elements_to_score: Vec<NodeRef<'_>> = vec![];
+            let doc = self.doc.clone();
+            let selection = doc.select_single("body");
+            let body_node = selection.nodes().first().unwrap();
+            let strip_unlikely = flags.contains(GrabFlags::StripUnlikelys);
+            filter_document(body_node, metadata, strip_unlikely);
+    
+            let descendants = body_node.descendants();
+    
+            for node in descendants.iter().filter(|n| n.is_element()) {
+                let Some(node_name) = node.node_name() else {
+                    unreachable!()
+                };
+    
+                if TAGS_WITH_CONTENT.contains(&node_name.as_ref()) {
+                    // TODO: this is a controversial moment, it may leave an empty block,
+                    // which will have an impact on the result.
+                    // When parent of the top candidate have more than one child,
+                    // then parent will be a new top candidate.
+    
+                    if is_element_without_content(node) {
+                        node.remove_from_parent();
+                        continue;
+                    }
+                }
+                // this block is relate to previous block
+                if node_name.as_ref() == "div" {
+                    div_into_p(node, &doc, &mut elements_to_score);
+                    continue;
+                }
+    
+                if DEFAULT_TAGS_TO_SCORE.contains(&node_name.as_ref()) {
+                    elements_to_score.push(node.clone());
+                }
+            }
+    
+            let article_node = self.handle_candidates(&mut elements_to_score, &doc, &flags);
+            let mut parse_successful = true;
+    
+            let mut article_doc: Option<Document> = None;
+    
+            if let Some(ref article_node) = article_node {
+                metadata.dir = get_dir_attr(article_node);
+                article_doc = Some(Document::from(article_node.html()));
+                let text_length = normalize_spaces(&article_node.text()).chars().count();
+                if text_length < DEFAULT_CHAR_THRESHOLD {
+                    parse_successful = false;
+    
+                    attempts.push((article_doc.clone(), text_length));
+                    if flags.contains(GrabFlags::StripUnlikelys) {
+                        flags -= GrabFlags::StripUnlikelys;
+                    } else if flags.contains(GrabFlags::WeightClasses) {
+                        flags -= GrabFlags::WeightClasses;
+                    } else if flags.contains(GrabFlags::CleanConditionally) {
+                        flags -= GrabFlags::CleanConditionally;
+                    } else {
+                        // No luck after removing flags, just return the longest text we found during the different loops
+                        attempts.sort_by_key(|i| Reverse(i.1));
+    
+                        if attempts[0].1 == 0 {
+                            return None;
+                        }
+                        article_doc = attempts[0].0.clone();
+                        parse_successful = true;
+                    }
+                }
+            } else {
+                parse_successful = false;
+            }
+    
+            if parse_successful {
+                return article_doc;
+            }
+            // Now that we've gone through the full algorithm, check to see if
+            // we got any meaningful content. If we didn't, we may need to re-run
+            // grabArticle with different flags set. This gives us a higher likelihood of
+            // finding the content, and the sieve approach gives us a higher likelihood of
+            // finding the -right- content.
+        }
+    }
+
+    fn handle_candidates<'a>(
+        &self,
+        elements_to_score: &mut Vec<NodeRef<'a>>,
+        doc: &'a Document,
+        flags: &FlagSet<GrabFlags>,
+    ) -> Option<NodeRef<'a>> {
+        let mut top_candidates = score_elements(elements_to_score, flags);
+        top_candidates.truncate(self.config.n_top_candidates);
+    
+        let mut top_candidate = top_candidates.first().cloned();
+    
+        let tc_name = top_candidate
+            .as_ref()
+            .and_then(|n| n.node_name())
+            .unwrap_or_else(StrTendril::new);
+    
+        let page_sel = doc.select("body");
+        let page_node = page_sel.nodes().first().unwrap();
+        let mut needed_to_create_top_candidate = false;
+    
+        if top_candidate.is_none() || tc_name.as_ref() == "body" {
+            needed_to_create_top_candidate = true;
+    
+            let tc = doc.tree.new_element("div");
+    
+            doc.tree.reparent_children_of(&page_node.id, Some(tc.id));
+            page_node.append_child(&tc);
+            init_node_score(&tc, flags.contains(GrabFlags::WeightClasses));
+            top_candidate = Some(tc);
+        } else if let Some(ref tc) = top_candidate {
+            // Find a better top candidate node if it contains (at least three) nodes which belong to `topCandidates` array
+            // and whose scores are quite closed with current `topCandidate` node.
+    
+            let tc_score = get_node_score(tc);
+    
+            let mut alternative_candidate_ancestors = vec![];
+            for alt in top_candidates.iter().skip(1) {
+                if get_node_score(alt) / tc_score >= 0.75 {
+                    alternative_candidate_ancestors.push(alt.ancestors(Some(0)));
+                }
+            }
+            if alternative_candidate_ancestors.len() > MINIMUM_TOP_CANDIDATES {
+                let mut parent_of_top_candidate = tc.parent();
+                while let Some(ref parent_of_tc) = parent_of_top_candidate {
+                    if node_name_is(parent_of_tc, "body") {
+                        break;
+                    }
+    
+                    let mut lists_containing_this_ancestor = 0;
+    
+                    for alt_ancestor in &alternative_candidate_ancestors {
+                        if lists_containing_this_ancestor >= MINIMUM_TOP_CANDIDATES {
+                            break;
+                        }
+    
+                        if alt_ancestor.iter().any(|n| n.id == parent_of_tc.id) {
+                            lists_containing_this_ancestor += 1;
+                        }
+                    }
+    
+                    if lists_containing_this_ancestor >= MINIMUM_TOP_CANDIDATES {
+                        top_candidate = parent_of_top_candidate;
+                        break;
+                    }
+    
+                    parent_of_top_candidate = parent_of_tc.parent();
+                }
+            }
+    
+            if let Some(ref tc) = top_candidate {
+                if !has_node_score(tc) {
+                    init_node_score(tc, flags.contains(GrabFlags::WeightClasses));
+                }
+                // Because of our bonus system, parents of candidates might have scores
+                // themselves. They get half of the node. There won't be nodes with higher
+                // scores than our topCandidate, but if we see the score going *up* in the first
+                // few steps up the tree, that's a decent sign that there might be more content
+                // lurking in other places that we want to unify in. The sibling stuff
+                // below does some of that - but only if we've looked high enough up the DOM
+                // tree.
+                let mut last_score = get_node_score(tc);
+                let score_threshold = last_score / 3.0;
+                let mut parent_of_top_candidate = tc.parent();
+                while let Some(ref parent_of_tc) = parent_of_top_candidate {
+                    if node_name_is(parent_of_tc, "body") {
+                        break;
+                    }
+    
+                    if !has_node_score(parent_of_tc) {
+                        parent_of_top_candidate = parent_of_tc.parent();
+                        continue;
+                    }
+    
+                    let parent_score = get_node_score(parent_of_tc);
+                    if parent_score < score_threshold {
+                        break;
+                    }
+                    if parent_score > last_score {
+                        top_candidate = parent_of_top_candidate;
+                        break;
+                    }
+                    last_score = parent_score;
+                    parent_of_top_candidate = parent_of_tc.parent();
+                }
+            }
+    
+            // If the top candidate is the only child, use parent instead. This will help sibling
+            // joining logic when adjacent content is actually located in parent's sibling node.
+            if let Some(ref tc) = top_candidate {
+                let mut parent_of_top_candidate = tc.parent();
+    
+                while let Some(ref parent_of_tc) = parent_of_top_candidate {
+                    if node_name_is(parent_of_tc, "body") {
+                        break;
+                    }
+    
+                    if parent_of_tc.element_children().len() != 1 {
+                        break;
+                    }
+                    top_candidate = parent_of_top_candidate.clone();
+                    parent_of_top_candidate = parent_of_tc.parent();
+                }
+            }
+        }
+        if let Some(ref tc) = top_candidate {
+            if !has_node_score(tc) {
+                init_node_score(tc, flags.contains(GrabFlags::WeightClasses));
+            }
+    
+            // Now that we have the top candidate, look through its siblings for content
+            // that might also be related. Things like preambles, content split by ads
+            // that we removed, etc.
+    
+            let mut article_content = doc.tree.new_element("div");
+            article_content.set_attr("id", "readability-content");
+    
+            handle_top_candidate(tc, &article_content);
+    
+            //prepare the article
+            prep_article(&article_content, flags);
+    
+            if needed_to_create_top_candidate {
+                // This looks like nonsense
+                // We already created a fake div thing, and there wouldn't have been any siblings left
+                // for the previous loop, so there's no point trying to create a new div, and then
+                // move all the children over. Just assign IDs and class names here. No need to append
+                // because that already happened anyway.
+                tc.set_attr("id", "readability-page-1");
+                tc.set_attr("class", "page");
+            } else {
+                let div = doc.tree.new_element("div");
+                div.set_attr("id", "readability-page-1");
+                div.set_attr("class", "page");
+                doc.tree
+                    .reparent_children_of(&article_content.id, Some(div.id));
+                article_content.replace_with(&div);
+                article_content = div;
+            }
+    
+            return Some(article_content);
+        }
+        None
+    }
+}
+
 
 fn filter_document(root_node: &NodeRef, metadata: &mut Metadata, strip_unlikely: bool) {
     let mut should_remove_title_header = !metadata.title.is_empty();
@@ -73,92 +328,7 @@ fn filter_document(root_node: &NodeRef, metadata: &mut Metadata, strip_unlikely:
     }
 }
 
-pub fn grab_article(doc: &Document, metadata: &mut Metadata) -> Option<Document> {
-    let mut flags =
-        GrabFlags::CleanConditionally | GrabFlags::StripUnlikelys | GrabFlags::WeightClasses;
 
-    let mut attempts = vec![];
-    loop {
-        let mut elements_to_score: Vec<NodeRef<'_>> = vec![];
-        let doc = doc.clone();
-        let selection = doc.select_single("body");
-        let body_node = selection.nodes().first().unwrap();
-        let strip_unlikely = flags.contains(GrabFlags::StripUnlikelys);
-        filter_document(body_node, metadata, strip_unlikely);
-
-        let descendants = body_node.descendants();
-
-        for node in descendants.iter().filter(|n| n.is_element()) {
-            let Some(node_name) = node.node_name() else {
-                unreachable!()
-            };
-
-            if TAGS_WITH_CONTENT.contains(&node_name.as_ref()) {
-                // TODO: this is a controversial moment, it may leave an empty block,
-                // which will have an impact on the result.
-                // When parent of the top candidate have more than one child,
-                // then parent will be a new top candidate.
-
-                if is_element_without_content(node) {
-                    node.remove_from_parent();
-                    continue;
-                }
-            }
-            // this block is relate to previous block
-            if node_name.as_ref() == "div" {
-                div_into_p(node, &doc, &mut elements_to_score);
-                continue;
-            }
-
-            if DEFAULT_TAGS_TO_SCORE.contains(&node_name.as_ref()) {
-                elements_to_score.push(node.clone());
-            }
-        }
-
-        let article_node = handle_candidates(&mut elements_to_score, &doc, &flags);
-        let mut parse_successful = true;
-
-        let mut article_doc: Option<Document> = None;
-
-        if let Some(ref article_node) = article_node {
-            metadata.dir = get_dir_attr(article_node);
-            article_doc = Some(Document::from(article_node.html()));
-            let text_length = normalize_spaces(&article_node.text()).chars().count();
-            if text_length < DEFAULT_CHAR_THRESHOLD {
-                parse_successful = false;
-
-                attempts.push((article_doc.clone(), text_length));
-                if flags.contains(GrabFlags::StripUnlikelys) {
-                    flags -= GrabFlags::StripUnlikelys;
-                } else if flags.contains(GrabFlags::WeightClasses) {
-                    flags -= GrabFlags::WeightClasses;
-                } else if flags.contains(GrabFlags::CleanConditionally) {
-                    flags -= GrabFlags::CleanConditionally;
-                } else {
-                    // No luck after removing flags, just return the longest text we found during the different loops
-                    attempts.sort_by_key(|i| Reverse(i.1));
-
-                    if attempts[0].1 == 0 {
-                        return None;
-                    }
-                    article_doc = attempts[0].0.clone();
-                    parse_successful = true;
-                }
-            }
-        } else {
-            parse_successful = false;
-        }
-
-        if parse_successful {
-            return article_doc;
-        }
-        // Now that we've gone through the full algorithm, check to see if
-        // we got any meaningful content. If we didn't, we may need to re-run
-        // grabArticle with different flags set. This gives us a higher likelihood of
-        // finding the content, and the sieve approach gives us a higher likelihood of
-        // finding the -right- content.
-    }
-}
 
 fn get_node_matching_string(node: &NodeRef) -> String {
     let mut matched_attrs: Vec<String> = vec![];
@@ -357,168 +527,7 @@ fn score_elements<'a>(
     candidates
 }
 
-fn handle_candidates<'a>(
-    elements_to_score: &mut Vec<NodeRef<'a>>,
-    doc: &'a Document,
-    flags: &FlagSet<GrabFlags>,
-) -> Option<NodeRef<'a>> {
-    let mut top_candidates = score_elements(elements_to_score, flags);
-    top_candidates.truncate(DEFAULT_N_TOP_CANDIDATES);
 
-    let mut top_candidate = top_candidates.first().cloned();
-
-    let tc_name = top_candidate
-        .as_ref()
-        .and_then(|n| n.node_name())
-        .unwrap_or_else(StrTendril::new);
-
-    let page_sel = doc.select("body");
-    let page_node = page_sel.nodes().first().unwrap();
-    let mut needed_to_create_top_candidate = false;
-
-    if top_candidate.is_none() || tc_name.as_ref() == "body" {
-        needed_to_create_top_candidate = true;
-
-        let tc = doc.tree.new_element("div");
-
-        doc.tree.reparent_children_of(&page_node.id, Some(tc.id));
-        page_node.append_child(&tc);
-        init_node_score(&tc, flags.contains(GrabFlags::WeightClasses));
-        top_candidate = Some(tc);
-    } else if let Some(ref tc) = top_candidate {
-        // Find a better top candidate node if it contains (at least three) nodes which belong to `topCandidates` array
-        // and whose scores are quite closed with current `topCandidate` node.
-
-        let tc_score = get_node_score(tc);
-
-        let mut alternative_candidate_ancestors = vec![];
-        for alt in top_candidates.iter().skip(1) {
-            if get_node_score(alt) / tc_score >= 0.75 {
-                alternative_candidate_ancestors.push(alt.ancestors(Some(0)));
-            }
-        }
-        if alternative_candidate_ancestors.len() > MINIMUM_TOP_CANDIDATES {
-            let mut parent_of_top_candidate = tc.parent();
-            while let Some(ref parent_of_tc) = parent_of_top_candidate {
-                if node_name_is(parent_of_tc, "body") {
-                    break;
-                }
-
-                let mut lists_containing_this_ancestor = 0;
-
-                for alt_ancestor in &alternative_candidate_ancestors {
-                    if lists_containing_this_ancestor >= MINIMUM_TOP_CANDIDATES {
-                        break;
-                    }
-
-                    if alt_ancestor.iter().any(|n| n.id == parent_of_tc.id) {
-                        lists_containing_this_ancestor += 1;
-                    }
-                }
-
-                if lists_containing_this_ancestor >= MINIMUM_TOP_CANDIDATES {
-                    top_candidate = parent_of_top_candidate;
-                    break;
-                }
-
-                parent_of_top_candidate = parent_of_tc.parent();
-            }
-        }
-
-        if let Some(ref tc) = top_candidate {
-            if !has_node_score(tc) {
-                init_node_score(tc, flags.contains(GrabFlags::WeightClasses));
-            }
-            // Because of our bonus system, parents of candidates might have scores
-            // themselves. They get half of the node. There won't be nodes with higher
-            // scores than our topCandidate, but if we see the score going *up* in the first
-            // few steps up the tree, that's a decent sign that there might be more content
-            // lurking in other places that we want to unify in. The sibling stuff
-            // below does some of that - but only if we've looked high enough up the DOM
-            // tree.
-            let mut last_score = get_node_score(tc);
-            let score_threshold = last_score / 3.0;
-            let mut parent_of_top_candidate = tc.parent();
-            while let Some(ref parent_of_tc) = parent_of_top_candidate {
-                if node_name_is(parent_of_tc, "body") {
-                    break;
-                }
-
-                if !has_node_score(parent_of_tc) {
-                    parent_of_top_candidate = parent_of_tc.parent();
-                    continue;
-                }
-
-                let parent_score = get_node_score(parent_of_tc);
-                if parent_score < score_threshold {
-                    break;
-                }
-                if parent_score > last_score {
-                    top_candidate = parent_of_top_candidate;
-                    break;
-                }
-                last_score = parent_score;
-                parent_of_top_candidate = parent_of_tc.parent();
-            }
-        }
-
-        // If the top candidate is the only child, use parent instead. This will help sibling
-        // joining logic when adjacent content is actually located in parent's sibling node.
-        if let Some(ref tc) = top_candidate {
-            let mut parent_of_top_candidate = tc.parent();
-
-            while let Some(ref parent_of_tc) = parent_of_top_candidate {
-                if node_name_is(parent_of_tc, "body") {
-                    break;
-                }
-
-                if parent_of_tc.element_children().len() != 1 {
-                    break;
-                }
-                top_candidate = parent_of_top_candidate.clone();
-                parent_of_top_candidate = parent_of_tc.parent();
-            }
-        }
-    }
-    if let Some(ref tc) = top_candidate {
-        if !has_node_score(tc) {
-            init_node_score(tc, flags.contains(GrabFlags::WeightClasses));
-        }
-
-        // Now that we have the top candidate, look through its siblings for content
-        // that might also be related. Things like preambles, content split by ads
-        // that we removed, etc.
-
-        let mut article_content = doc.tree.new_element("div");
-        article_content.set_attr("id", "readability-content");
-
-        handle_top_candidate(tc, &article_content);
-
-        //prepare the article
-        prep_article(&article_content, flags);
-
-        if needed_to_create_top_candidate {
-            // This looks like nonsense
-            // We already created a fake div thing, and there wouldn't have been any siblings left
-            // for the previous loop, so there's no point trying to create a new div, and then
-            // move all the children over. Just assign IDs and class names here. No need to append
-            // because that already happened anyway.
-            tc.set_attr("id", "readability-page-1");
-            tc.set_attr("class", "page");
-        } else {
-            let div = doc.tree.new_element("div");
-            div.set_attr("id", "readability-page-1");
-            div.set_attr("class", "page");
-            doc.tree
-                .reparent_children_of(&article_content.id, Some(div.id));
-            article_content.replace_with(&div);
-            article_content = div;
-        }
-
-        return Some(article_content);
-    }
-    None
-}
 
 fn handle_top_candidate(tc: &Node, article_content: &Node) {
     let tc_node_score = get_node_score(tc);
@@ -668,13 +677,14 @@ mod tests {
                  <h6></h6>
             </body>
         </html>"#;
-
-        let doc = Document::from(contents);
-        let sel = doc.select("body > *");
+        
+        let ra = Readability::new(contents, None, None).unwrap();
+        let sel = ra.doc.select("body > *");
         let count_before = sel.nodes().iter().filter(|n| n.is_element()).count();
 
         assert_eq!(count_before, 10);
-        let clean_doc = grab_article(&doc, &mut Metadata::default()).unwrap();
+
+        let clean_doc = ra.grab_article(&mut Metadata::default()).unwrap();
         let sel = clean_doc.select("body > *");
         let count_after = sel.nodes().iter().filter(|n| n.is_element()).count();
         assert_eq!(count_after, 1);
